+++ cppad/local/optimize/record_vp.hpp
@@ -83,7 +83,7 @@ struct_size_pair record_vp(
 	CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < npar  );
 	//
 	addr_t new_arg[2];
-	new_arg[0]   = new_var[ random_itr.var2op(arg[0]) ];
+	new_arg[0]   = new_var[ random_itr.var2op(size_t(arg[0])) ];
 	new_arg[1]   = new_par[ arg[1] ];
 	rec->PutArg( new_arg[0], new_arg[1] );
 	//
